backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
google-generativeai
openai
python-multipart
python-dotenv
alembic
pydantic-settings
